add useImportedItems hook
and use it in QueryResutls.tsx
